participant soft delete

+++ src/db/repositories/ParticipantRepo.js
@@ -7,28 +7,29 @@ class ParticipantRepo {
         return {
             id: participant.id,
             cage: participant.cage,
-            image: participant.image,
-            image_flipped: participant.image_flipped,
-            team_id: participant.team_id,
-            event_id: participant.event_id,
-            stadium_id: participant.stadium_id,
-            stadium_name: participant.stadium_name,
-            betting_pref: participant.betting_pref,
-            betting_amount: participant.betting_amount,
-            owner_account_number: participant.owner_account_number,
-            type: participant.type,
-            color: participant.color,
-            cresta: participant.cresta,
-            alas: participant.alas,
-            pata: participant.pata,
-            breeder_id: participant.breeder_id,
-            breeder_name: participant.breeder_name,
-            weight: participant.weight,
-            participated_before: participant.participated_before,
-            physical_advantage: participant.physical_advantage,
+            image: participant.image || null,
+            image_flipped: participant.image_flipped || false,
+            team_id: participant.team_id || null,
+            event_id: participant.event_id || null,
+            stadium_id: participant.stadium_id || null,
+            stadium_name: participant.stadium_name || null,
+            betting_pref: participant.betting_pref || null,
+            betting_amount: participant.betting_amount || null,
+            owner_account_number: participant.owner_account_number || null,
+            type: participant.type || null,
+            color: participant.color || null,
+            cresta: participant.cresta || null,
+            alas: participant.alas || null,
+            pata: participant.pata || null,
+            breeder_id: participant.breeder_id || null,
+            breeder_name: participant.breeder_name || null,
+            weight: participant.weight || null,
+            participated_before: participant.participated_before || null,
+            physical_advantage: participant.physical_advantage || null,
             status: participant.status,
             reason: participant.reason,
-            observation: participant.observation,
+            observation: participant.observation || null,
+            deleted: participant.deleted || false
         }
     }
 
@@ -38,7 +39,7 @@ class ParticipantRepo {
 
     async getApprovedByEventId(event_id) {
         return await participants.findAll({
-            where: {event_id, status: "approved"},
+            where: {event_id, status: "approved", deleted: false},
             include: [
                 {model:teams, include: [{model: team_owners, include: [{model: mutual_liberty, as: 'owner_liberty'}]}]}
             ],
@@ -48,19 +49,53 @@ class ParticipantRepo {
 
     async upsertParticipant(participant) {
         if (participant.id) {
-            await participants.update(participant, {where: {id: participant.id}});
+            await participants.update({...participant, deleted: false}, {where: {id: participant.id}});
             return participant;
         } else {
-            return await participants.create(participant);
+            const lastParticipant = await participants.findOne({where: {event_id: participant.event_id}, order: [['cage', 'DESC']]});
+            const newCage = lastParticipant ? lastParticipant.cage + 1 : 1;
+            return await participants.create({...participant, cage: newCage});
         }
     }
 
     async removeParticipant(id) {
-        return await participants.destroy({where: {id}});
+        const participant = await participants.findOne({where: {id}});
+        if (participant.deleted) {
+            const deleted = await participants.destroy({where: {id}});
+            await this.updateParticipantCages(participant.event_id);
+            return deleted;
+        } else if (participant) {
+            return await participants.update(this.toParticipant({
+                id: participant.id,
+                event_id: participant.event_id,
+                cage: participant.cage,
+                status: participant.status,
+                reason: participant.reason,
+                deleted: true
+            }), {where: {id}});
+        }
     }
 
     async findParticipantByPayload(payload) {
         return await participants.findOne({where: payload});
+    }
+
+    async removeDeleted(event_id) {
+        const participantsToDelete = await participants.findAll({where: {event_id, deleted: true}});
+        const participantsIds = participantsToDelete.map((m) => m.id);
+        if (!participantsIds) return false;
+
+        return await participants.destroy({where: {id: participantsIds} });
+    }
+
+    async updateParticipantCages(event_id) {
+        const allEventParticipants = await participants.findAll({where: {event_id}, order: [['cage', 'ASC']]});
+        allEventParticipants.map((p, i) => {
+            p.cage = i + 1;
+            p.save();
+        });
+
+        return true;
     }
 }
